pyproject.toml
[project]
name = "sticky"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-cors-headers>=4.6.0",
    "django>=5.1.4",
    "djangorestframework>=3.15.2",
    "nanoid>=2.0.0",
    "pydantic>=2.10.3",
]

[tool.pyright]
venv = "."
venvDir = "."
